Debug Strings functionality

+++ Application/Commands/Strings/Delete/DeleteStringCommandHandler.cs
@@ -1,4 +1,5 @@
-﻿using Application.Interfaces;
+﻿using Application.Common.Exceptions;
+using Application.Interfaces;
 using AspNetCore.Yandex.ObjectStorage;
 using Microsoft.EntityFrameworkCore;
 
@@ -11,6 +12,10 @@ namespace Application.Commands.Strings.Delete
         public async override Task Handle(DeleteStringCommand request, CancellationToken cancellationToken) // TODO: Обработка исключений
         {
             var dbStr = await _dbContext.Strings.FirstOrDefaultAsync(str => str.Code == str.Code, cancellationToken);
+            if (dbStr is null)
+            {
+                throw new NotFoundException(nameof(Domain.String), request.Code);
+            }
             _dbContext.Strings.Remove(dbStr);
             await _dbContext.SaveChangesAsync(cancellationToken);
         }
